feat!: использование конфига для работы с проектам
refactor: рефакторинг проекта
