+++ lib/package/templates/additionalFiles.ts
