+++ lib/package/templates/packageIcon.ts
