--- lib/templates/widget/src/widgetAppDTs.ts
@@ -1,67 +1,0 @@
-import { WIDGET_SDK_LIB_NAME } from "../../../const.js";
-
-export const APP_D_TS_TEMPLATE = `\
-/* eslint-disable @typescript-eslint/triple-slash-reference */
-/// <reference types="@infomaximum/global-types" />
-/// <reference types="${WIDGET_SDK_LIB_NAME}" />
-
-declare module "*.gif" {
-  const src: string;
-  export default src;
-}
-
-declare module "*.jpg" {
-  const src: string;
-  export default src;
-}
-
-declare module "*.jpeg" {
-  const src: string;
-  export default src;
-}
-
-declare module "*.png" {
-  const src: string;
-  export default src;
-}
-
-declare module "*.webp" {
-  const src: string;
-  export default src;
-}
-
-declare module "*.svg" {
-  import * as React from "react";
-
-  export const ReactComponent: React.FunctionComponent<
-    React.SVGProps<SVGSVGElement> & { title?: string }
-  >;
-
-  export default ReactComponent;
-}
-
-declare module "*.svg?url" {
-  const src: string;
-  export default src;
-}
-
-declare module "*.module.css" {
-  const classes: { readonly [key: string]: string };
-  export default classes;
-}
-
-declare module "*.module.scss" {
-  const classes: { readonly [key: string]: string };
-  export default classes;
-}
-
-declare module "*.module.sass" {
-  const classes: { readonly [key: string]: string };
-  export default classes;
-}
-
-declare module "*.module.less" {
-  const classes: { readonly [key: string]: string };
-  export default classes;
-}
-`;
