+++ lib/utils.ts
@@ -4,13 +4,12 @@ import path from "node:path";
 import { spawn, type SpawnOptions, exec } from "node:child_process";
 import util from "node:util";
 import Module from "node:module";
-import semver from "semver";
-import chalk from "chalk";
+
 import {
   CUSTOM_PACKAGE_CLI_LIB_NAME,
   MANIFEST_SERVICE_FIELDS_FOR_DEVELOPMENT,
-  WIDGET_SDK_LIB_NAME,
 } from "./const.js";
+import { WIDGET_SDK_LIB_NAME } from "./widget/const.js";
 
 const execPromise = util.promisify(exec);
 
@@ -92,53 +91,11 @@ export function spawnCommand(
   });
 }
 
-export async function checkLatestVersion(libName: string) {
-  const libVersionInProject = await getLibraryVersionInProject(libName);
-
-  if (!libVersionInProject) return;
-
-  const libVersionFromProject: string | undefined = (
-    libVersionInProject?.dependencies?.[libName] ||
-    libVersionInProject?.devDependencies?.[libName]
-  )?.version;
-
-  if (!libVersionFromProject) return;
-
-  const latestVersion = await getLatestVersionOfLibrary(libName);
-
-  if (!latestVersion) return;
-
-  const isOldVersion = semver.gt(latestVersion, libVersionFromProject);
-
-  if (isOldVersion) {
-    console.error(
-      chalk.yellow(
-        `\n\nA new version of the ${chalk.underline(
-          `${libName}@${latestVersion}`
-        )} library has been released,\n` +
-          `old version ${libVersionFromProject} is used in the project, it is recommended to ` +
-          `update to the latest version \n` +
-          chalk.green.bold(
-            chalk.underline(`npm i --save ${libName}@${latestVersion}`) +
-              " or " +
-              chalk.underline(`yarn upgrade ${libName}@${latestVersion}\n\n`)
-          )
-      )
-    );
-  }
-}
-
-export async function checkLatestLibsVersion() {
-  try {
-    await Promise.allSettled([
-      checkLatestVersion(WIDGET_SDK_LIB_NAME),
-      checkLatestVersion(CUSTOM_PACKAGE_CLI_LIB_NAME),
-    ]);
-  } catch (error) {}
-}
-
 export function removeServiceFieldsForDevelopment(obj: Record<string, any>) {
   MANIFEST_SERVICE_FIELDS_FOR_DEVELOPMENT.forEach((key) => {
     delete obj[key];
   });
 }
+
+export const compact = <T>(items: (T | null | undefined | false | "" | 0)[]) =>
+  items.filter(Boolean) as T[];
