+++ lib/widget/configs/webpack/sections/loaders/cssLoaders.ts
@@ -1,0 +1,17 @@
+import { systemRequire } from "../../../../../utils.js";
+
+export const cssLoaders = [
+  systemRequire.resolve("style-loader"),
+  systemRequire.resolve("css-loader"),
+  {
+    loader: systemRequire.resolve("postcss-loader"),
+    options: {
+      postcssOptions: {
+        plugins: [
+          systemRequire.resolve("postcss-preset-env"),
+          systemRequire.resolve("autoprefixer"),
+        ],
+      },
+    },
+  },
+];
