+++ lib/widget/configs/webpack/sections/plugins/minimizer.ts
