+++ lib/widget/configs/webpack/sections/plugins/reactRefresh.ts
