+++ lib/widget/const.ts
@@ -1,0 +1,9 @@
+import { BUILD_ARCHIVE_EXT } from "../const.js";
+
+export const WIDGET_SDK_LIB_NAME = "@infomaximum/widget-sdk";
+export const WIDGET_CONFIG_FIELD_NAME = "widget";
+export const WIDGET_CONFIG_FILE_NAME = `${WIDGET_CONFIG_FIELD_NAME}rc`;
+export const WIDGET_ARCHIVE_NAME = "widget";
+export const WIDGET_ARCHIVE_FULL_NAME = `${WIDGET_ARCHIVE_NAME}.${BUILD_ARCHIVE_EXT}`;
+export const WIDGET_DEFAULT_HOST = "localhost";
+export const WIDGET_DEFAULT_PORT = "5555";
