+++ lib/widget/index.ts
@@ -1,0 +1,1 @@
+export { registerWidgetCommands } from "./commands.js";
