+++ lib/widget/scripts/init/generators.ts
