+++ lib/widget/scripts/init/helpers.ts
