+++ lib/widget/scripts/init/prompts.ts
