+++ lib/widget/templates/src/widgetIndexCSS.ts
