+++ lib/widget/templates/widgetConfigs.ts
