+++ package.json
@@ -24,7 +24,7 @@
   "dependencies": {
     "@babel/core": "^7.23.6",
     "@babel/plugin-transform-runtime": "^7.23.6",
-    "@infomaximum/assert": "1.1.2",
+    "@infomaximum/assert": "^1.1.3",
     "@infomaximum/json-modify-webpack-plugin": "^1.0.2",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.5.11",
     "@svgr/webpack": "^8.1.0",
@@ -43,6 +43,7 @@
     "postcss": "^8.4.32",
     "postcss-loader": "^7.3.3",
     "postcss-preset-env": "^9.3.0",
+    "rc-config-loader": "^4.1.3",
     "react-refresh": "^0.14.0",
     "remove-files-webpack-plugin": "^1.5.0",
     "sass": "^1.69.5",
