+++ schemas/widgetConfigSchema.json
@@ -1,0 +1,40 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "type": "object",
+  "properties": {
+    "entry": {
+      "type": "string"
+    },
+    "widgetManifest": {
+      "type": "string"
+    },
+    "packageManifest": {
+      "type": "string"
+    },
+    "packageDir": {
+      "type": "string"
+    },
+    "assetsDir": {
+      "type": "string"
+    },
+    "buildDir": {
+      "type": "string"
+    },
+    "port": {
+      "type": "integer",
+      "minimum": 0,
+      "maximum": 65535
+    },
+    "host": {
+      "type": "string"
+    }
+  },
+  "required": [
+    "entry",
+    "widgetManifest",
+    "packageManifest",
+    "packageDir",
+    "port",
+    "host"
+  ]
+}
